# services/gateway/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_fieldstone_fieldstone_services_records_handlers.listResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      requests:
        items:
          $ref: '#/definitions/handlers.FOIARequestResponse'
        type: array
      total:
        type: integer
    type: object
...
  handlers.FOIARequestResponse:
    properties:
      closed_at:
        type: string
      created_at:
        type: string
      department_id:
        type: string
      description:
        type: string
      due_date:
        description: '"2006-01-02" or null'
        type: string
      id:
        type: string
      metadata:
        items:
          type: integer
        type: array
      requester_email:
        type: string
      requester_name:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  handlers.createFOIARequestInput:
    properties:
      department_id:
        type: string
      description:
        type: string
      due_date:
        description: optional "2006-01-02"
        type: string
      metadata:
        items:
          type: integer
        type: array
      requester_email:
        type: string
      requester_name:
        type: string
    type: object
...
  handlers.updateStatusInput:
    properties:
      role:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Fieldstone Contributors
    url: https://github.com/fieldstone/fieldstone
  description: |-
    A self-hostable platform for city governments to run common civic services.

    Staff endpoints require a Bearer JWT from the configured OIDC provider.
    Citizen endpoints (marked "public") do not require authentication.

    **Development**: set `DEV_DISABLE_AUTH=true` in .env to bypass JWT
    validation locally. Remove before production deployment.
  license:
    name: MIT
    url: https://github.com/fieldstone/fieldstone/blob/main/LICENSE
  title: Fieldstone Civic Services API
  version: 0.1.0
paths:
...
  /v1/records/foia:
    get:
      parameters:
      - default: 20
        description: Max results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Pagination offset
        in: query
        name: offset
        type: integer
      - description: Filter by status (received, processing, awaiting_clarification,
          fulfilled, denied, withdrawn)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fieldstone_fieldstone_services_records_handlers.listResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List FOIA requests (staff)
      tags:
      - records
    post:
      consumes:
      - application/json
      description: Public endpoint — no authentication required.
      parameters:
      - description: FOIA request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createFOIARequestInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.FOIARequestResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Metadata validation failed
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Submit a FOIA request (public)
      tags:
      - records
  /v1/records/foia/{id}:
    get:
      parameters:
      - description: FOIA request UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.FOIARequestResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a FOIA request
      tags:
      - records
  /v1/records/foia/{id}/status:
    patch:
      consumes:
      - application/json
      description: |-
        Validates the transition via the workflow service. Terminal statuses
        (fulfilled, denied, withdrawn) also set closed_at.
      parameters:
      - description: FOIA request UUID
        in: path
        name: id
        required: true
        type: string
      - description: Status transition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.updateStatusInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.FOIARequestResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Transition not allowed by workflow
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update FOIA request status
      tags:
      - records
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token: `Bearer <token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"